+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## Unreleased
 
 ## 0.1.0
+ - port from private repo
+ - based folder and testing structure on that used by the Pester project
+ - adopted "manifest and changelog" and "style rules" policy from Pester project
+ - one function is implemented: Expand-String
 
 This changelog is inspired by the 
 [Pester](https://github.com/pester/Pester/blob/master/CHANGELOG.md) file